requirements.txt
argon2-cffi~=25.1.0
psycopg2~=2.9.10
psutil~=7.0.0
python-dotenv~=1.1.1